pyproject.toml
[project]
name = "merge-conflict-env"
version = "0.1.0"
description = "OpenEnv RL environment for merge conflict resolution"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic>=2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
